Fixed backup when no config file present.

+++ Scripts/steam/backup_steam.sh
@@ -4,7 +4,12 @@
 ls ~/config/ > ~/.config.txt
 
 # Sync steam
-rsync -ar --delete --exclude-from=~/.config.txt ~/local/.steam ~/backup
+if [ -s ~/.config.txt ]
+    then
+        rsync -ar --delete ~/local/.steam ~/backup
+    else
+        rsync -ar --delete --exclude-from=~/.config.txt ~/local/.steam ~/backup
+fi
 rsync -ar --delete ~/local/Steam ~/backup
 rsync -ar --delete ~/local/arma3 ~/backup
